init: First commit to new repository
This is the first commit of the project onto this newly-created repo.
Previously, the application was committed to an older repository that
had to be deprecated for security reasons. This commit is therefore
complete with rich features and functionality.

The previous repository that this application belonged to contained
sensitive information that needed to be removed. Rather than filter
through the commit history and leaving convoluted commit messages, a
clean commit to a new repository was decidedly a better option for
this project.